requirements.txt
# Core dependencies
anthropic>=0.39.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Date/Time
python-dateutil>=2.8.2

# Markdown generation
jinja2>=3.1.2
